+++ src/main/groovy/net/eikehirsch/gradle/react/JSXExtension.groovy
@@ -1,0 +1,12 @@
+package net.eikehirsch.gradle.react
+/**
+ * Extension to configure the jsx transpiler
+ */
+class JSXExtension {
+
+  public static String NAME = 'jsx'
+
+  String sourcesDir;
+  String destDir;
+
+}