+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Version 0.2 *(not released)*
 ----------------------------
 
 * Update to gradle 2.2
-* ...
+* sourcesDir and destDir can be configured via `jsx` extension or via extra task definition.
 
 Version 0.1 *(2014-11-12)*
 ----------------------------